Small upgrade to the menu for PC
Title + it still needsa fix for scrolling.
When the menu is opened on mobile the site beneath should not scroll, but while on pc, it should scroll.
It seems that the mobile unscrolling is ready but it needs a fix for when the menus are changed, Like from normal the quick one is opened. I have no idea why I'm writing this, no one reads it anyway....